Cleanup: use uint8_t for various flags in curves
Previously, it was kind of a mess. In different places it was using `char`, `short` and `int`.

The changed properties are flags that are operated upon using bit operations. Therefore, the integer type should be unsigned. Since we only use 2 bits of these flags, `uint8_t` is large enough.

Especially note the change I had to make in `RNA_define.h` to make this work.

Reviewers: campbellbarton

Differential Revision: https://developer.blender.org/D8844

+++ source/blender/blenkernel/BKE_lattice.h
@@ -77,7 +77,7 @@ int BKE_lattice_index_flip(
     struct Lattice *lt, const int index, const bool flip_u, const bool flip_v, const bool flip_w);
 void BKE_lattice_bitmap_from_flag(struct Lattice *lt,
                                   unsigned int *bitmap,
-                                  const short flag,
+                                  const uint8_t flag,
                                   const bool clear,
                                   const bool respecthide);
