+++ source/blender/editors/curve/editcurve_select.c
@@ -57,7 +57,7 @@
 #include "DEG_depsgraph.h"
 
 /* returns 1 in case (de)selection was successful */
-bool select_beztriple(BezTriple *bezt, bool selstatus, short flag, eVisible_Types hidden)
+bool select_beztriple(BezTriple *bezt, bool selstatus, uint8_t flag, eVisible_Types hidden)
 {
   if ((bezt->hide == 0) || (hidden == HIDDEN)) {
     if (selstatus == SELECT) { /* selects */
@@ -77,7 +77,7 @@ bool select_beztriple(BezTriple *bezt, bool selstatus, short flag, eVisible_Type
 }
 
 /* returns 1 in case (de)selection was successful */
-bool select_bpoint(BPoint *bp, bool selstatus, short flag, bool hidden)
+bool select_bpoint(BPoint *bp, bool selstatus, uint8_t flag, bool hidden)
 {
   if ((bp->hide == 0) || (hidden == 1)) {
     if (selstatus == SELECT) {
